Refactor message and behavior handling into separate managers for modularity
